modify the naming convention used when creating DLML figures
We used to use the CSV filename, but all files are called fragility.csv
and consequence_repair.csv in the new DLML. So, instead, now we are
using the path to the file in the DLML. This identifies the method well.
When the user provides their own data, we still use the filename as
that seems more useful.